--- app/views/helloworld/hello_world.scala.html
@@ -1,5 +1,0 @@
-@()(implicit request: Request[_], messages: Messages))
-
-@views.html.main_template(title = "Hello from cgt-agent-client-relationships-frontend", bodyClasses = None) {
-    <h1>Hello from cgt-agent-client-relationships-frontend !</h1>
-}
