+++ app/views/home/home.scala.html
@@ -1,0 +1,5 @@
+@()(implicit request: Request[_], messages: Messages)
+
+@views.html.main_template(title = "Lifetime ISA", bodyClasses = None) {
+    <h1>Title goes here!</h1>
+}
