+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             controllers.HelloWorld.helloWorld
+GET        /             controllers.HomePage.home
